% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmsma-package.R
\docType{package}
\name{bmsma-package}
\alias{bmsma-package}
\alias{bmsma}
\title{The 'bmsma' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R package version 2.26.23. https://mc-stan.org
}
