% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmsma_assign_data.R
\name{bmsma_assign_data}
\alias{bmsma_assign_data}
\title{Assign data to template}
\usage{
bmsma_assign_data(model_template, ...)
}
\arguments{
\item{model_template}{output from bmsma_model}

\item{...}{data-masking name-value pairs}
}
\value{
updated named list with your data assigned to Stan model parameters
}
\description{
Assign data to template
}
\examples{
bmsma_model("linear") |> bmsma_assign_data(X = Loblolly$age, Y = Loblolly$height)
}
