% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmot_lm.R
\name{rmot_lm}
\alias{rmot_lm}
\title{Run a linear model in Stan}
\usage{
rmot_lm(x, y, ...)
}
\arguments{
\item{x}{vector of numeric or integer predictor variable}

\item{y}{vector of numeric or integer response variable}

\item{...}{additional arguments passed to rstan::sampling}
}
\value{
Stan model output
}
\description{
Run a linear model in Stan
}
\examples{
mtcars
rmot_lm(mtcars$mpg, mtcars$disp)
}
