% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmsma_models.R
\name{bmsma_model}
\alias{bmsma_model}
\title{Select data configuration template for bmsma supported model}
\usage{
bmsma_model(model = NULL)
}
\arguments{
\item{model}{model name character string}
}
\value{
named list that matches Stan model parameters
}
\description{
Select data configuration template for bmsma supported model
}
\examples{
bmsma_model("linear")
}
