% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmsma_run.R
\name{bmsma_run}
\alias{bmsma_run}
\title{Run a linear model in Stan}
\usage{
bmsma_run(model_template, ...)
}
\arguments{
\item{model_template}{model template generated by bmsma_model and updated by bmsma_assign_data}

\item{...}{additional arguments passed to rstan::sampling}
}
\value{
Stanfit model output
}
\description{
Run a linear model in Stan
}
\examples{
bmsma_model("linear") |>
  bmsma_assign_data(X = Loblolly$age,
                   Y = Loblolly$height,
                   N = nrow(Loblolly)) |>
  bmsma_run()
}
