% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmot-package.R
\docType{package}
\name{rmot-package}
\alias{rmot-package}
\alias{rmot}
\title{The 'rmot' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R package version 2.26.23. https://mc-stan.org
}
