% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmot_assign_data.R
\name{rmot_assign_data}
\alias{rmot_assign_data}
\title{Assign data to template}
\usage{
rmot_assign_data(model_template, ...)
}
\arguments{
\item{model_template}{output from rmot_model}

\item{...}{data-masking name-value pairs}
}
\value{
updated named list with your data assigned to Stan model parameters
}
\description{
Assign data to template
}
\examples{
rmot_model("linear") |> rmot_assign_data(X = Loblolly$age, Y = Loblolly$height)
}
