% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmot_models.R
\name{rmot_model}
\alias{rmot_model}
\title{Select data configuration template for rmot supported model}
\usage{
rmot_model(model = NULL)
}
\arguments{
\item{model}{model name character string}
}
\value{
named list that matches Stan model parameters
}
\description{
Select data configuration template for rmot supported model
}
\examples{
rmot_model("linear")
}
